update: refactor dmg system.

+++ Assets/Scripts/Entities/IDamageable.cs
@@ -4,6 +4,6 @@ namespace Entities
 {
     public interface IDamageable
     {
-        void TakeDamage(int damage);
+        void TakeDamage(float damage);
     }
 }
